chore(Sections): rename renderer -> view
makes more sense
