+++ packages/wix-storybook-utils/src/Sections/index.test.ts
@@ -5,20 +5,20 @@ import * as builders from './index';
 
 import { SectionType } from '../typings/story-section';
 
-const cwd = path.resolve(__dirname, 'renderers');
+const cwd = path.resolve(__dirname, 'views');
 const methodToFileName = f => kebabCase(path.parse(f).name);
 
 const sectionTypes = Object.keys(SectionType).map(t => SectionType[t]);
 
 describe('Sections', () => {
-  it('should be exported in renderers', () => {
+  it('should have exported view for each section', () => {
     sectionTypes.map(type => {
       try {
-        const renderer = require(path.resolve(cwd, methodToFileName(type)));
-        expect(typeof renderer[type]).toBe('function');
+        const view = require(path.resolve(cwd, methodToFileName(type)));
+        expect(typeof view[type]).toBe('function');
       } catch (e) {
         throw new Error(
-          `Missing renderer function for "${type}" section type. Make sure one exists in src/Sections/renderers. ERROR: ${e}`,
+          `Missing view function for "${type}" section type. Make sure one exists in src/Sections/views. ERROR: ${e}`,
         );
       }
     });
